\name{dataobjects}
\docType{data}

\alias{dataobjects}

\alias{sum1}
\alias{sum2}
\alias{sum3}

\alias{sum1f}
\alias{sum2f}
\alias{sum3f}

\alias{sum1M}
\alias{sum2M}
\alias{sum3M}

\alias{sum1MT}
\alias{sum2MT}
\alias{sum3MT}

\alias{sum1MCL}
\alias{sum2MCL}
\alias{sum3MCL}

\alias{sum2CCL}
\alias{sum2CfCL}

\title{ Data Objects }

\description{

Simulation summaries as generated in the vignette.

}

\usage{

sum1
sum2
sum3

sum1f
sum2f
sum3f

sum1M
sum2M
sum3M

sum1MT
sum2MT
sum3MT

sum1MCL
sum2MCL
sum3MCL

sum2CCL
sum2CfCL

}

\seealso{

\link{summary}

}
