\name{overdispsim-package}
\alias{overdispsim}

\docType{package}

\title{Simulations of Overdispersion in SECR}

\description{

Code used to simulate overdispersion for paper by Efford and Fletcher.
  
}

\details{

\tabular{ll}{

Package: \tab overdispsim\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab 2025-05-19\cr
License: \tab GNU General Public License Version 2 or later\cr

}

Type \code{vignette('Overview', 'overdispsim')} to see the overview vignette.

}

\seealso{

\code{\link{run_all}}

}
