\name{addCL}
\alias{addCL}

\title{Add Confidence Interval to Summary Plot}

\description{

Intended for plot of summary results.

}

\usage{

addCL(x, summ, cov = "COV", nrepl = 1000, ...)

}

\arguments{
  \item{x}{numeric x values}
  \item{summ}{dataframe from summary_M etc.}
  \item{cov}{character statistic to plot e.g. RB}
  \item{nrepl}{integer number of replicates for binomial interval for COV statistic}
  \item{\dots}{other arguments passed to Bucl, Blcl}
}


\details{

Nothing yet.

}

\value{

None returned

}

\seealso{

\code{\link{summary_M}}

}
